refactor: split app into seperate component files
